+++ app/consumer-democracy/proposals_whitelisting_test.go
@@ -3,9 +3,9 @@ package app_test
 import (
 	"testing"
 
-	appConsumer "github.com/cosmos/interchain-security/v2/app/consumer-democracy"
-	ibctesting "github.com/cosmos/interchain-security/v2/legacy_ibc_testing/testing"
-	icstestingutils "github.com/cosmos/interchain-security/v2/testutil/ibc_testing"
+	appConsumer "github.com/cosmos/interchain-security/v3/app/consumer-democracy"
+	ibctesting "github.com/cosmos/interchain-security/v3/legacy_ibc_testing/testing"
+	icstestingutils "github.com/cosmos/interchain-security/v3/testutil/ibc_testing"
 	"github.com/stretchr/testify/require"
 )
 
@@ -13,7 +13,7 @@ func TestDemocracyGovernanceWhitelistingKeys(t *testing.T) {
 	chain := ibctesting.NewTestChain(t, ibctesting.NewCoordinator(t, 0),
 		icstestingutils.DemocracyConsumerAppIniter, "test")
 	paramKeeper := chain.App.(*appConsumer.App).ParamsKeeper
-	for paramKey := range appConsumer.WhitelistedParams {
+	for paramKey := range appConsumer.LegacyWhitelistedParams {
 		ss, ok := paramKeeper.GetSubspace(paramKey.Subspace)
 		require.True(t, ok, "Unknown subspace %s", paramKey.Subspace)
 		hasKey := ss.Has(chain.GetContext(), []byte(paramKey.Key))