+++ app/consumer/ante/disabled_modules_ante_test.go
@@ -7,9 +7,9 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	evidencetypes "github.com/cosmos/cosmos-sdk/x/evidence/types"
 	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
-	ibcclienttypes "github.com/cosmos/ibc-go/v4/modules/core/02-client/types"
-	"github.com/cosmos/interchain-security/v2/app/consumer/ante"
-	"github.com/cosmos/interchain-security/v2/app/params"
+	ibcclienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
+	"github.com/cosmos/interchain-security/v3/app/consumer/ante"
+	"github.com/cosmos/interchain-security/v3/app/params"
 	"github.com/stretchr/testify/require"
 )