+++ x/ccv/provider/proposal_handler_test.go
@@ -10,12 +10,13 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 
-	clienttypes "github.com/cosmos/ibc-go/v4/modules/core/02-client/types"
+	clienttypes "github.com/cosmos/ibc-go/v7/modules/core/02-client/types"
 
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
-	testkeeper "github.com/cosmos/interchain-security/v2/testutil/keeper"
-	"github.com/cosmos/interchain-security/v2/x/ccv/provider"
-	providertypes "github.com/cosmos/interchain-security/v2/x/ccv/provider/types"
+	govv1beta1 "github.com/cosmos/cosmos-sdk/x/gov/types/v1beta1"
+
+	testkeeper "github.com/cosmos/interchain-security/v3/testutil/keeper"
+	"github.com/cosmos/interchain-security/v3/x/ccv/provider"
+	providertypes "github.com/cosmos/interchain-security/v3/x/ccv/provider/types"
 )
 
 // TestProviderProposalHandler tests the highest level handler for proposals
@@ -28,7 +29,7 @@ func TestProviderProposalHandler(t *testing.T) {
 
 	testCases := []struct {
 		name                     string
-		content                  govtypes.Content
+		content                  govv1beta1.Content
 		blockTime                time.Time
 		expValidConsumerAddition bool
 		expValidConsumerRemoval  bool
@@ -79,9 +80,13 @@ func TestProviderProposalHandler(t *testing.T) {
 		},
 		{
 			name: "unsupported proposal type",
-			content: distributiontypes.NewCommunityPoolSpendProposal(
-				"title", "desc", []byte{},
-				sdk.NewCoins(sdk.NewCoin("communityfunds", sdk.NewInt(10)))),
+			// lint rule disabled because this is a test case for an unsupported proposal type
+			content: &distributiontypes.CommunityPoolSpendProposal{ // nolint:staticcheck
+				Title:       "title",
+				Description: "desc",
+				Recipient:   "",
+				Amount:      sdk.NewCoins(sdk.NewCoin("communityfunds", sdk.NewInt(10))),
+			},
 		},
 	}